+++ src/Messaging.Core/EnvelopePublisher.cs
@@ -12,9 +12,14 @@ public sealed class EnvelopePublisher<TPayload> : IEnvelopePublisher<TPayload>
         this.publicationPipeline = publicationPipeline ?? throw new ArgumentNullException(nameof(publicationPipeline));
     }
 
-    public RawEnvelope Publish(TypedEnvelope<TPayload> source, Func<TPayload, string> payloadSerializer)
+    public RawEnvelope PublishInline(TypedEnvelope<TPayload> source, Func<TPayload, string> payloadSerializer)
     {
         return publicationPipeline.Compose(source, payloadSerializer);
+    }
+
+    public RawEnvelope Publish(TypedEnvelope<TPayload> source, Func<TPayload, string> payloadSerializer)
+    {
+        return PublishInline(source, payloadSerializer);
     }
 
     public RawEnvelope PublishWithReference(